backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
email-validator
